moved EmbeddingsParams to api dir
